Update Malawi2020Election.py
This software prompts the user to enter total number of published centers, total number of registered voters, total number of null and void votes, total number of valid votes and total valid votes for each candidate. Afterward, it determines the candidate with the majority of votes and displays the results on the screen.
NOTE: For a candidate to be a majority winner the candidate total valid votes should be greater than majority votes.